[ade7880] split ade7880 into base and i2c
