+++ CODEOWNERS
@@ -19,6 +19,8 @@ esphome/components/adc/* @esphome/core
 esphome/components/adc128s102/* @DeerMaximum
 esphome/components/addressable_light/* @justfalter
 esphome/components/ade7880/* @kpfleming
+esphome/components/ade7880_base/* @cyberhuman @kpfleming
+esphome/components/ade7880_i2c/* @kpfleming
 esphome/components/ade7953/* @angelnu
 esphome/components/ade7953_i2c/* @angelnu
 esphome/components/ade7953_spi/* @angelnu
