+++ esphome/components/ade7880_i2c/__init__.py
@@ -1,0 +1,1 @@
+CODEOWNERS = ["@kpfleming"]
