+++ tests/components/ade7880_i2c/test.esp32-ard.yaml
