+++ tests/components/ade7880_i2c/test.esp32-c3-ard.yaml
