+++ tests/components/ade7880_i2c/test.esp8266-ard.yaml
