+++ tests/components/ade7880_i2c/test.rp2040-ard.yaml
